+++ rocprim/include/rocprim/device/device_reduce_by_key.hpp
@@ -25,6 +25,7 @@
 #include "device_reduce_by_key_config.hpp"
 #include "device_transform.hpp"
 
+#include "detail/device_config_helper.hpp"
 #include "detail/device_reduce_by_key.hpp"
 #include "detail/device_scan_common.hpp"
 #include "detail/lookback_scan_state.hpp"
@@ -47,11 +48,8 @@ BEGIN_ROCPRIM_NAMESPACE
 namespace detail
 {
 
-namespace reduce_by_key
-{
-
 template<typename LookBackScanState, typename AccumulatorType>
-ROCPRIM_KERNEL __launch_bounds__(ROCPRIM_DEFAULT_MAX_BLOCK_SIZE) void init_kernel(
+ROCPRIM_KERNEL __launch_bounds__(ROCPRIM_DEFAULT_MAX_BLOCK_SIZE) void reduce_by_key_init_kernel(
     LookBackScanState              lookback_scan_state,
     const unsigned int             number_of_tiles,
     ordered_block_id<unsigned int> ordered_bid,
@@ -102,22 +100,23 @@ template<typename Config,
          typename CompareFunction,
          typename BinaryOp,
          typename LookbackScanState>
-ROCPRIM_KERNEL __launch_bounds__(Config::block_size) void kernel(
-    const KeyIterator                    keys_input,
-    const ValueIterator                  values_input,
-    const UniqueIterator                 unique_keys,
-    const ReductionIterator              reductions,
-    const UniqueCountIterator            unique_count,
-    const BinaryOp                       reduce_op,
-    const CompareFunction                compare,
-    const LookbackScanState              scan_state,
-    const ordered_block_id<unsigned int> ordered_tile_id,
-    const std::size_t                    starting_tile,
-    const std::size_t                    total_number_of_tiles,
-    const std::size_t                    size,
-    const std::size_t* const             global_head_count,
-    const AccumulatorType* const         previous_accumulated,
-    const std::size_t                    number_of_tiles_launch)
+ROCPRIM_KERNEL
+    __launch_bounds__(device_params<Config>().kernel_config.block_size) void reduce_by_key_kernel(
+        const KeyIterator                    keys_input,
+        const ValueIterator                  values_input,
+        const UniqueIterator                 unique_keys,
+        const ReductionIterator              reductions,
+        const UniqueCountIterator            unique_count,
+        const BinaryOp                       reduce_op,
+        const CompareFunction                compare,
+        const LookbackScanState              scan_state,
+        const ordered_block_id<unsigned int> ordered_tile_id,
+        const std::size_t                    starting_tile,
+        const std::size_t                    total_number_of_tiles,
+        const std::size_t                    size,
+        const std::size_t* const             global_head_count,
+        const AccumulatorType* const         previous_accumulated,
+        const std::size_t                    number_of_tiles_launch)
 {
     reduce_by_key::kernel_impl<Config>(keys_input,
                                        values_input,
@@ -179,9 +178,15 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
     using key_type         = reduce_by_key::value_type_t<KeysInputIterator>;
     using accumulator_type = reduce_by_key::accumulator_type_t<ValuesInputIterator, BinaryFunction>;
 
-    using config = detail::default_or_custom_config<
-        Config,
-        reduce_by_key::default_config<ROCPRIM_TARGET_ARCH, key_type, accumulator_type>>;
+    using config = wrapped_reduce_by_key_config<Config, key_type, accumulator_type, BinaryFunction>;
+
+    detail::target_arch target_arch;
+    hipError_t          result = host_target_arch(stream, target_arch);
+    if(result != hipSuccess)
+    {
+        return result;
+    }
+    const reduce_by_key_config_params params = dispatch_target_arch<config>(target_arch);
 
     using scan_state_type
         = reduce_by_key::lookback_scan_state_t<accumulator_type, /*UseSleep=*/false>;
@@ -190,13 +195,13 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
 
     using ordered_tile_id_type = detail::ordered_block_id<unsigned int>;
 
-    constexpr unsigned int block_size      = config::block_size;
-    constexpr unsigned int tiles_per_block = config::tiles_per_block;
-    constexpr unsigned int items_per_tile  = block_size * config::items_per_thread;
-    constexpr unsigned int items_per_block = items_per_tile * tiles_per_block;
+    const unsigned int block_size      = params.kernel_config.block_size;
+    const unsigned int tiles_per_block = params.tiles_per_block;
+    const unsigned int items_per_tile  = block_size * params.kernel_config.items_per_thread;
+    const unsigned int items_per_block = items_per_tile * tiles_per_block;
 
-    static constexpr size_t size_limit = config::size_limit;
-    static constexpr size_t aligned_size_limit
+    const size_t size_limit = params.kernel_config.size_limit;
+    const size_t aligned_size_limit
         = ::rocprim::max<size_t>(size_limit - size_limit % items_per_block, items_per_block);
 
     const size_t limited_size     = std::min<size_t>(size, aligned_size_limit);
@@ -215,14 +220,13 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
     accumulator_type* d_previous_accumulated = nullptr;
 
     detail::temp_storage::layout layout{};
-    const hipError_t             layout_result
-        = scan_state_type::get_temp_storage_layout(number_of_tiles, stream, layout);
-    if(layout_result != hipSuccess)
+    result = scan_state_type::get_temp_storage_layout(number_of_tiles, stream, layout);
+    if(result != hipSuccess)
     {
-        return layout_result;
+        return result;
     }
 
-    const hipError_t partition_result = detail::temp_storage::partition(
+    result = detail::temp_storage::partition(
         temporary_storage,
         storage_size,
         detail::temp_storage::make_linear_partition(
@@ -233,30 +237,31 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
             detail::temp_storage::ptr_aligned_array(&d_global_head_count, use_limited_size ? 1 : 0),
             detail::temp_storage::ptr_aligned_array(&d_previous_accumulated,
                                                     use_limited_size ? 1 : 0)));
-    if(partition_result != hipSuccess || temporary_storage == nullptr)
-    {
-        return partition_result;
-    }
-
-    bool             use_sleep;
-    const hipError_t result = detail::is_sleep_scan_state_used(stream, use_sleep);
-    if(result != hipSuccess)
+    if(result != hipSuccess || temporary_storage == nullptr)
     {
         return result;
     }
 
-    scan_state_type scan_state{};
-    hipError_t      scan_state_result
-        = scan_state_type::create(scan_state, scan_state_storage, number_of_tiles, stream);
-    scan_state_with_sleep_type scan_state_with_sleep{};
-    scan_state_result = scan_state_with_sleep_type::create(scan_state_with_sleep,
-                                                           scan_state_storage,
-                                                           number_of_tiles,
-                                                           stream);
-
-    if(scan_state_result != hipSuccess)
+    bool use_sleep;
+    result = detail::is_sleep_scan_state_used(stream, use_sleep);
+    if(result != hipSuccess)
     {
-        return scan_state_result;
+        return result;
+    }
+    scan_state_type scan_state{};
+    result = scan_state_type::create(scan_state, scan_state_storage, number_of_tiles, stream);
+    if(result != hipSuccess)
+    {
+        return result;
+    }
+    scan_state_with_sleep_type scan_state_with_sleep{};
+    result = scan_state_with_sleep_type::create(scan_state_with_sleep,
+                                                scan_state_storage,
+                                                number_of_tiles,
+                                                stream);
+    if(result != hipSuccess)
+    {
+        return result;
     }
 
     auto with_scan_state
@@ -308,7 +313,6 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
         const std::size_t number_of_tiles_launch = ceiling_div(current_size, items_per_tile);
         const std::size_t number_of_blocks_launch
             = ceiling_div(number_of_tiles_launch, tiles_per_block);
-        const std::size_t init_grid_size = detail::ceiling_div(number_of_tiles_launch, block_size);
 
         // Start point for time measurements
         std::chrono::high_resolution_clock::time_point start;
@@ -325,18 +329,18 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
         with_scan_state(
             [&](const auto scan_state)
             {
-                hipLaunchKernelGGL(init_kernel,
-                                   dim3(init_grid_size),
-                                   dim3(block_size),
-                                   0,
-                                   stream,
-                                   scan_state,
-                                   number_of_tiles_launch,
-                                   ordered_bid,
-                                   i == 0,
-                                   number_of_tiles - 1,
-                                   d_global_head_count,
-                                   d_previous_accumulated);
+                const unsigned int block_size = ROCPRIM_DEFAULT_MAX_BLOCK_SIZE;
+                const std::size_t  grid_size
+                    = detail::ceiling_div(number_of_tiles_launch, block_size);
+
+                reduce_by_key_init_kernel<<<dim3(grid_size), dim3(block_size), 0, stream>>>(
+                    scan_state,
+                    number_of_tiles_launch,
+                    ordered_bid,
+                    i == 0,
+                    number_of_tiles - 1,
+                    d_global_head_count,
+                    d_previous_accumulated);
             });
         ROCPRIM_DETAIL_HIP_SYNC_AND_RETURN_ON_ERROR("init_lookback_scan_state_kernel",
                                                     number_of_tiles_launch,
@@ -345,26 +349,23 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
         with_scan_state(
             [&](const auto scan_state)
             {
-                hipLaunchKernelGGL(reduce_by_key::kernel<config>,
-                                   dim3(number_of_blocks_launch),
-                                   dim3(block_size),
-                                   0,
-                                   stream,
-                                   keys_input + offset,
-                                   values_input + offset,
-                                   unique_output,
-                                   aggregates_output,
-                                   unique_count_output,
-                                   reduce_op,
-                                   key_compare_op,
-                                   scan_state,
-                                   ordered_bid,
-                                   i * number_of_tiles,
-                                   total_number_of_tiles,
-                                   size,
-                                   i > 0 ? d_global_head_count : nullptr,
-                                   i > 0 ? d_previous_accumulated : nullptr,
-                                   number_of_tiles_launch);
+                reduce_by_key_kernel<config>
+                    <<<dim3(number_of_blocks_launch), dim3(block_size), 0, stream>>>(
+                        keys_input + offset,
+                        values_input + offset,
+                        unique_output,
+                        aggregates_output,
+                        unique_count_output,
+                        reduce_op,
+                        key_compare_op,
+                        scan_state,
+                        ordered_bid,
+                        i * number_of_tiles,
+                        total_number_of_tiles,
+                        size,
+                        i > 0 ? d_global_head_count : nullptr,
+                        i > 0 ? d_previous_accumulated : nullptr,
+                        number_of_tiles_launch);
             });
         ROCPRIM_DETAIL_HIP_SYNC_AND_RETURN_ON_ERROR("reduce_by_key_kernel", current_size, start);
     }
@@ -373,8 +374,6 @@ hipError_t reduce_by_key_impl(void*                     temporary_storage,
 }
 
 #undef ROCPRIM_DETAIL_HIP_SYNC_AND_RETURN_ON_ERROR
-
-} // namespace reduce_by_key
 
 } // namespace detail
 
@@ -506,18 +505,18 @@ inline hipError_t reduce_by_key(void*                     temporary_storage,
                                 hipStream_t               stream            = 0,
                                 bool                      debug_synchronous = false)
 {
-    return detail::reduce_by_key::reduce_by_key_impl<Config>(temporary_storage,
-                                                             storage_size,
-                                                             keys_input,
-                                                             values_input,
-                                                             size,
-                                                             unique_output,
-                                                             aggregates_output,
-                                                             unique_count_output,
-                                                             reduce_op,
-                                                             key_compare_op,
-                                                             stream,
-                                                             debug_synchronous);
+    return detail::reduce_by_key_impl<Config>(temporary_storage,
+                                              storage_size,
+                                              keys_input,
+                                              values_input,
+                                              size,
+                                              unique_output,
+                                              aggregates_output,
+                                              unique_count_output,
+                                              reduce_op,
+                                              key_compare_op,
+                                              stream,
+                                              debug_synchronous);
 }
 
 /// @}